Fixed Instantaneous Factors
Instantaneous factors fixed. The reason why it was not working was that the temporal variables that would be part of the instantaneous factors were placed below in the code segment. Thus, the instantaneous factors had a value of 0 for the temporal variables. By rearranging the equations in the code segment, I was able to fix #16 and close #12